requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/dev.in
#
-e file:///home/stuart/Development/django-feeds
    # via -r requirements/dev.in
alabaster==0.7.13
    # via sphinx
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via pytest
babel==2.12.1
    # via sphinx
billiard==3.6.4.0
    # via celery
black==23.1.0
    # via -r requirements/tests.in
bleach==6.0.0
    # via readme-renderer
bump2version==1.0.1
    # via -r requirements/dev.in
cachetools==5.3.0
    # via tox
celery==5.2.7
    # via django-feeds
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
chardet==5.1.0
    # via tox
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via tox
colorclass==2.2.2
    # via pip-upgrader
coverage[toml]==7.2.1
    # via pytest-cov
croniter==1.3.8
    # via django-feeds
cryptography==39.0.2
    # via secretstorage
distlib==0.3.6
    # via virtualenv
django==4.1.7
    # via
    #   django-debug-toolbar
    #   django-extensions
    #   django-feeds
    #   django-stubs
    #   django-stubs-ext
    #   django-tagulous
django-debug-toolbar==3.8.1
    # via -r requirements/dev.in
django-extensions==3.2.1
    # via django-feeds
django-stubs==1.15.0
    # via -r requirements/dev.in
django-stubs-ext==0.7.0
    # via django-stubs
django-tagulous==1.3.3
    # via django-feeds
docopt==0.6.2
    # via pip-upgrader
docutils==0.18.1
    # via
    #   readme-renderer
    #   sphinx
    #   sphinx-rtd-theme
exceptiongroup==1.1.0
    # via pytest
factory-boy==3.2.1
    # via pytest-factoryboy
faker==17.6.0
    # via factory-boy
feedparser==6.0.10
    # via django-feeds
filelock==3.9.0
    # via
    #   tox
    #   virtualenv
flake8==6.0.0
    # via -r requirements/tests.in
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==6.0.0
    # via
    #   keyring
    #   twine
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements/tests.in
jaraco-classes==3.2.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via sphinx
keyring==23.13.1
    # via twine
kombu==5.2.4
    # via celery
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
more-itertools==9.1.0
    # via jaraco-classes
mypy==1.1.1
    # via
    #   -r requirements/tests.in
    #   django-stubs
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.0
    # via
    #   black
    #   pip-upgrader
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
pathspec==0.11.0
    # via black
pip-upgrader==1.4.15
    # via -r requirements/dev.in
pkginfo==1.9.6
    # via twine
platformdirs==3.1.1
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
prompt-toolkit==3.0.38
    # via click-repl
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.14.0
    # via
    #   readme-renderer
    #   rich
    #   sphinx
pyproject-api==1.5.0
    # via tox
pytest==7.2.2
    # via
    #   -r requirements/tests.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
pytest-cov==4.0.0
    # via -r requirements/tests.in
pytest-django==4.5.2
    # via -r requirements/tests.in
pytest-factoryboy==2.5.1
    # via -r requirements/tests.in
python-dateutil==2.8.2
    # via
    #   croniter
    #   django-feeds
    #   faker
pytz==2022.7.1
    # via celery
readme-renderer==37.3
    # via twine
requests==2.28.2
    # via
    #   django-feeds
    #   pip-upgrader
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==0.10.1
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.3.2
    # via twine
secretstorage==3.3.3
    # via keyring
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   bleach
    #   click-repl
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==6.1.3
    # via
    #   -r requirements/docs.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.2.0
    # via -r requirements/docs.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jquery==2.0.0
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
terminaltables==3.1.10
    # via pip-upgrader
text-unidecode==1.3
    # via django-feeds
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   django-stubs
    #   mypy
    #   pyproject-api
    #   pytest
    #   tox
tox==4.4.6
    # via -r requirements/dev.in
twine==4.0.2
    # via -r requirements/dev.in
types-pytz==2022.7.1.2
    # via django-stubs
types-pyyaml==6.0.12.8
    # via django-stubs
typing-extensions==4.5.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pytest-factoryboy
urllib3==1.26.14
    # via
    #   requests
    #   twine
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.20.0
    # via tox
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
